requirements.txt
Flask
qrcode
pynacl